requirements.txt
# Core FastAPI and web server dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# HTTP and async client libraries
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Google API dependencies
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Calendar and iCal parsing
icalendar>=5.0.0

# Web scraping and parsing
beautifulsoup4==4.13.5
lxml==4.9.3
feedparser==6.0.10

# Email security and DNS verification (FounderShield)
dnspython>=2.4.0
python-whois>=0.8.0

# Desktop application
pywebview>=4.4.1

# Data analysis and visualization
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# AI and machine learning
openai==1.3.0

# Voice system (TTS and speech recognition)
# Note: Voice listening requires PyAudio, which needs portaudio19-dev
# Install with: sudo apt install portaudio19-dev python3-pyaudio
# Then: pip install SpeechRecognition PyAudio
# Voice output (TTS) works without PyAudio

# Task scheduling and monitoring
schedule==1.2.0
psutil==5.9.6

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
